refactor: move code from main to src/app.js
This change will facilitate testing of the code in the main module. This
way I'll be able to mock downstream calls and test printing of errors to
stderr and process exit status.

Also reworked the code a bit to use promisified pipeline() call.
